requirements.txt
streamlit
google-cloud-bigquery
google-generativeai 
Pillow 
requests 
vertexai 
json 
streamlit-option-menu 
time 
hashlib 
google.generativeai 
Pillow 
Image 
vertexai